+++ script/strategies/kandel/deployers/MumbaiKandelSeederDeployer.s.sol
@@ -34,7 +34,8 @@ contract MumbaiKandelSeederDeployer is Deployer {
       deployERC4626Kandel: true,
       deployKandel: true,
       testBase: IERC20(fork.get("CRV.T/AAVEv3")), // make sure to use an AAVE faucet
-      testQuote: IERC20(fork.get("WBTC.T/AAVEv3")) // make sure to use an AAVE faucet
+      testQuote: IERC20(fork.get("WBTC.T/AAVEv3")), // make sure to use an AAVE faucet,
+      wnative: address(0)
     });
 
     // Pass the struct to innerRun