+++ script/strategies/kandel/deployers/PolygonKandelSeederDeployer.s.sol
@@ -34,7 +34,8 @@ contract PolygonKandelSeederDeployer is Deployer {
       deployERC4626Kandel: true,
       deployKandel: true,
       testBase: IERC20(fork.get("WETH.e")),
-      testQuote: IERC20(fork.get("DAI.e"))
+      testQuote: IERC20(fork.get("DAI.e")),
+      wnative: address(0)
     });
 
     // Pass the struct to innerRun