+++ test/strategies/kandel/AaveKandel.t.sol
@@ -66,7 +66,8 @@ contract AaveKandelTest is CoreKandelTest {
       mgv: IMangrove($(mgv)),
       olKeyBaseQuote: olKey,
       gasreq: kandel_gasreq,
-      routerParams: Direct.RouterParams({routerImplementation: router, fundOwner: id, strict: strict})
+      routerParams: Direct.RouterParams({routerImplementation: router, fundOwner: id, strict: strict}),
+      wnative_: address(0)
     });
 
     router.bind(address(aaveKandel_));
@@ -368,7 +369,8 @@ contract AaveKandelTest is CoreKandelTest {
       mgv: IMangrove($(mgv)),
       gasreq: 700_000,
       olKeyBaseQuote: OLKey(address(aToken), address(quote), 1),
-      routerParams: Direct.RouterParams({routerImplementation: router, fundOwner: address(0), strict: false})
+      routerParams: Direct.RouterParams({routerImplementation: router, fundOwner: address(0), strict: false}),
+      wnative_: address(0)
     });
   }
 
@@ -380,7 +382,8 @@ contract AaveKandelTest is CoreKandelTest {
       mgv: IMangrove($(mgv)),
       gasreq: 700_000,
       olKeyBaseQuote: OLKey(address(base), address(aToken), 1),
-      routerParams: Direct.RouterParams({routerImplementation: router, fundOwner: address(0), strict: false})
+      routerParams: Direct.RouterParams({routerImplementation: router, fundOwner: address(0), strict: false}),
+      wnative_: address(0)
     });
   }
 }