+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Change licenses to MIT for code mistakenly assigned BSD-2-Clause or AGPL-3.0
 - Fix submodule configurations for mangrove-core, v2-core, v3-core, and v3-periphery
--
+- Added ERC4626 RouterProxy deployer
 
 # 2.1.0-9